requirements.txt
tqdm==4.61.2
numpy==1.24.2
torch==2.4.1
torchvision==0.14.1
omegaconf==2.3.0
matplotlib==3.7.1
Pillow==9.4.0
huggingface_hub==0.26.5
transformers==4.46.3
accelerate==1.0.1
diffusers==0.31.0
tokenizers==0.20.3
sentencepiece==0.1.99
ftfy==6.1.3
beautifulsoup4==4.11.2
protobuf==3.19.6
gradio==4.44.1